Add NPS retrieval method and mock data for testing
This commit introduces a new method, get_nps, in the ConversationsMetricsService to retrieve the Net Promoter Score (NPS) for a project. It also adds mock data for NPS metrics to facilitate testing. Corresponding unit tests are implemented to validate the functionality of the get_nps method, ensuring accurate retrieval of NPS values.